Move the tested examples to example_test.go
Makes the actual test files a bit easier to manage IMHO, and because we
can use "package toml_test" it also means that the examples can use
"toml.Decode" instead of just "Decode", which is a bit clearer in the
docs and allows for easier copy/pasting.
